いくつかの関数をdatabase.tsに統合
